fix: checkout GITHUB_REF when performing a pull request event
GITHUB_REF is the branch on the origin repository containing the pull request.
GITHUB_HEAD_REF is the REMOTE branch on the FORK that corresponds to GITHUB_REF (what the contributor submitted as the pull request branch)
GITHUB_BASE_REF is the branch on the origin repository against which the PR was submitted.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>