fix: run as non-privileged user
We occasionally have test cases that are checking for the ability to do things like write to non-writeable files.

The standard image runs as the root user, which is needed to ensure we can install extensions, update configuration, set the preferred PHP version, etc.

However, the root user can _always_ write to files.

There are tools, such as `chattr +i <filename>` that can mark a file immutable, but these are **disabled in Docker**.

As such, the solution I came up with is to:

- Add a "testuser" account
- Add the "sudo" package
- In the entrypoint:
  - change ownership of the workdir and its contents to testuser
  - run the command as testuser

This resolves the testing issues, while retaining the original functionality.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>