feat: make checkout more flexible
When testing locally, we can explicitly set the workdir to the current checkout, which means no checkout is really required.
As such, we can omit the various `$GITHUB_*` env variables when invoking the container, and simplify usage.
The first conditional in the updated `checkout()` function does this for us, and the conditional within the `pull_request` case acts as an extra safeguard.
Example of such usage:

```bash
docker run -v $(realpath .):/github/workspace -w=/github/workspace laminas-check-runner:latest '{"php":"7.4","deps":"latest","extensions":[],"ini":["memory_limit=-1"],"command":"./vendor/bin/phpunit"}'
```

When testing as part of a local action, where we might want to change the entrypoint, we need to first run `actions/checkout`, which mans we already have a checkout; performing a clone into the workdir thus raises an error.
To make this usage work, I added the `if [ -d ".git" ];then` conditional, which adds the "origin" remote and points it at the canonical repository (removing the remote first if it already exists), and the fetches it so we get history.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>