feat: install sqlsrv and pdo_sqlsrv extensions for PHP 7.3-8.0
The 5.9.0 version of the MS sqlsrv extensions provides support for Ubuntu 20.04, for PHP versions 7.3, 7.4, and 8.0.
The require the msodbcsql17 libraries from the MS DEB repository, but are themselves shipped as binaries via GitHub.
This patch adds the MS DEB repo to the container, installs msodbcsql17, and pulls the sqlsrv and pdo-sqlsrv extensions for each of PHP 7.3, 7.4, and 8.0 from GitHub, installing them in the appropriate extension library directory for each version.
It adds configuration to the `mods-available` directory of each PHP version.
The `entrypoint.sh` is altered to look for any references to sqlsrv extensions, and, when discovered:

- Emits a notice on PHP versions < 7.3 indicating it cannot be enabled
- Uses `phpenmod` to enable the sqlsrv configuration for the requested PHP version

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>